requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altair=4.1.0=pypi_0
argon2-cffi=20.1.0=pypi_0
astor=0.8.1=pypi_0
attrs=20.1.0=pypi_0
backcall=0.2.0=pypi_0
base58=2.0.1=pypi_0
bleach=3.1.5=pypi_0
blinker=1.4=pypi_0
boto3=1.14.47=pypi_0
botocore=1.17.47=pypi_0
ca-certificates=2020.6.24=0
cachetools=4.1.1=pypi_0
certifi=2020.6.20=py38_0
cffi=1.14.2=pypi_0
chardet=3.0.4=pypi_0
click=7.1.2=pypi_0
colorama=0.4.3=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.6.0=pypi_0
docutils=0.15.2=pypi_0
entrypoints=0.3=pypi_0
enum-compat=0.0.3=pypi_0
idna=2.10=pypi_0
ipykernel=5.3.4=pypi_0
ipython=7.17.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
jedi=0.17.2=pypi_0
jinja2=2.11.2=pypi_0
jmespath=0.10.0=pypi_0
joblib=0.16.0=pypi_0
jsonschema=3.2.0=pypi_0
jupyter-client=6.1.6=pypi_0
jupyter-core=4.6.3=pypi_0
markupsafe=1.1.1=pypi_0
mistune=0.8.4=pypi_0
nbconvert=5.6.1=pypi_0
nbformat=5.0.7=pypi_0
notebook=6.1.3=pypi_0
numpy=1.19.1=pypi_0
openssl=1.1.1g=he774522_1
packaging=20.4=pypi_0
pandas=1.1.1=pypi_0
pandocfilters=1.4.2=pypi_0
parso=0.7.1=pypi_0
pathtools=0.1.2=pypi_0
pickleshare=0.7.5=pypi_0
pillow=7.2.0=pypi_0
pip=20.2.2=py38_0
prometheus-client=0.8.0=pypi_0
prompt-toolkit=3.0.6=pypi_0
protobuf=3.13.0=pypi_0
pyarrow=1.0.1=pypi_0
pycparser=2.20=pypi_0
pydeck=0.4.1=pypi_0
pygments=2.6.1=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.16.0=pypi_0
python=3.8.5=he1778fa_0
python-dateutil=2.8.1=pypi_0
pytz=2020.1=pypi_0
pywin32=228=pypi_0
pywinpty=0.5.7=pypi_0
pyzmq=19.0.2=pypi_0
requests=2.24.0=pypi_0
s3transfer=0.3.3=pypi_0
scikit-learn=0.23.1=pypi_0
scipy=1.5.2=pypi_0
send2trash=1.5.0=pypi_0
setuptools=49.6.0=py38_0
six=1.15.0=pypi_0
sqlite=3.33.0=h2a8f88b_0
streamlit=0.65.0=pypi_0
terminado=0.8.3=pypi_0
testpath=0.4.4=pypi_0
threadpoolctl=2.1.0=pypi_0
toml=0.10.1=pypi_0
toolz=0.10.0=pypi_0
tornado=6.0.4=pypi_0
traitlets=4.3.3=pypi_0
tzlocal=2.1=pypi_0
urllib3=1.25.10=pypi_0
validators=0.18.0=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
watchdog=0.10.3=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.34.2=py38_0
widgetsnbextension=3.5.1=pypi_0
wincertstore=0.2=py38_0
zlib=1.2.11=h62dcd97_4
